+++ src/main/java/com/ds/ds/domain/user/util/UserConverter.java
@@ -2,11 +2,13 @@ package com.ds.ds.domain.user.util;
 
 import com.ds.ds.domain.group.domain.entity.Group;
 import com.ds.ds.domain.user.domain.entity.User;
+import com.ds.ds.domain.user.presentation.data.dto.HeaderDto;
 import com.ds.ds.domain.user.presentation.data.dto.UpdateUserNameDto;
 import com.ds.ds.domain.user.presentation.data.dto.UpdateUserProfileImgDto;
 import com.ds.ds.domain.user.presentation.data.dto.UserDto;
 import com.ds.ds.domain.user.presentation.data.request.UpdateUserNameRequest;
 import com.ds.ds.domain.user.presentation.data.request.UpdateUserProfileImgRequest;
+import com.ds.ds.domain.user.presentation.data.response.HeaderResponse;
 import com.ds.ds.domain.user.presentation.data.response.UserResponse;
 
 import java.util.List;
@@ -16,7 +18,8 @@ public interface UserConverter {
     UserDto.GroupDto toDto(Group group);
     UpdateUserNameDto toDto(UpdateUserNameRequest request);
     UpdateUserProfileImgDto toDto(UpdateUserProfileImgRequest request);
-
+    HeaderDto toDto(User user);
     UserResponse toResponse(UserDto dto, List<UserResponse.GroupResponse> responses);
     UserResponse.GroupResponse toResponse(UserDto.GroupDto dto);
+    HeaderResponse toResponse(HeaderDto headerDto);
 }
